+++ src/main/kotlin/cloud/glitchdev/rfu/config/categories/GeneralFishing.kt
@@ -71,9 +71,9 @@ object GeneralFishing : Category("General Fishing") {
         reloadScreen()
     }
 
-    var rareDropMessageFormat by string("&6&lRARE DROP! &e{drop} &b(+{magic_find}% ✯ Magic Find) &7(Took {count} catches, {time} since last)") {
+    var rareDropMessageFormat by string("&6&lRARE DROP! &e{drop} &b(+{magic_find} ✯ Magic Find) &7(Took {count} catches, {time} since last)") {
         name = Literal("Custom Message Format")
-        description = Literal("Variables: {drop}, {magic_find}, {count}, {time}")
+        description = Literal("Variables: {drop}, {magic_find}, {count}, {time}, {total}")
         condition = { customRareDropMessage }
     }
 
@@ -87,6 +87,27 @@ object GeneralFishing : Category("General Fishing") {
         name = Literal("Lootshare Message")
         description = Literal("Sends a message when lootshare gives you an item.")
     }
+
+    var rareDropTitleAlert by observable(boolean(true) {
+        name = Literal("Rare Drop Title Alert")
+        description = Literal("Shows a title on screen when you get a rare drop")
+    }) { _, _ ->
+        reloadScreen()
+    }
+
+        var rareDropTitleFormat by string("{dropcolor}&l{drop} &f#{total}") {
+            name = Literal("Rare Drop Title Format")
+            description = Literal("The title to show on screen. Variables: {drop}, {dropcolor}, {magic_find}, {count}, {time}, {total}")
+            condition = { rareDropTitleAlert }
+        }
+
+        var rareDropSubtitleFormat by string("&b(+{magic_find} ✯ Magic Find)") {
+            name = Literal("Rare Drop Subtitle Format")
+            description = Literal("The subtitle to show on screen. Variables: {drop}, {dropcolor}, {magic_find}, {count}, {time}, {total}")
+            condition = { rareDropTitleAlert }
+        }
+
+
 
     init {
         dualSeparator {